fix: bad implementation of the inactivity feature
The problem is that the inactivity feature is that it's checking for the player.queue.current which either returns 0 if theres only one in queue, or it return null when theres nothing in the queue.
The new implementation of the inactivity feature is that It will check for the player.playing instead. :p

Also prettyfied the code.

Signed-off-by: Darren <git@darrennathanael.com>
Reviewed-by: Darren <git@darrennathanael.com>