feat: reply on ping.
When the bot is pinged it will respond with a message.
--
My prefix on this server is / (Slash Command).
To get started you can type /help to see all my commands.
If you can't see it, Please reinvite me with the correct permissions.
--

This is so that people know that the bot is slash commands only.
I don't know if it will works after April 30, 2022. due to discord api changes.

Signed-off-by: Darren <git@darrennathanael.com>
Signed-off-by: dottodottoo <90232327+dottodottoo@users.noreply.github.com>
Reviewed-by: Darren <git@darrennathanael.com>